eeg_analysis/requirements.txt
# Core dependencies
numpy>=1.22.4
pandas>=1.3.5
scipy>=1.10.0
scikit-learn>=1.0.2

# Signal processing
nolds>=0.5.2

# ML frameworks
mlflow>=2.8.0
imbalanced-learn>=0.10.1

# Data handling
pyarrow>=14.0.1
h5py>=3.8.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2

# Notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Development tools
black>=23.7.0
pylint>=2.17.5
pytest>=7.4.0
pytest-cov>=4.1.0

# Configuration
pyyaml>=6.0.1

# Logging
python-json-logger>=2.0.7